Don't use y-position of circles to style circles
Using the y-position of the circles to style them causes Forced
Synchronous Layout because you have to query the DOM.

Since we have a indexable container of story elements, calculate how far
down the page as a percentage using the total number of elements and the
index of the current element and use this percentage to style the
circles. This calculation doesn't touch the DOM, so Forced Synchronous
Layout won't happen.