Optimize scrolling for reduced jank
Idea here is to have the score circles at a good size in the intial CSS
and only scale a subset of them with JavaScript. The circles were often
too small anyway toward the bottom of the page. Now they're set in the
intial CSS to be about the right size for a three-digit score, and are
never scaled down.