Display correct article age
Display correct article age on the slider that comes out when an
article is clicked.

The solution here was to just multiply the article age by 1000 because
it looks like the Hacker News API provides the timestamps in seconds,
whereas JavaScript works in milliseconds.

To fix this properly, one would to this conversion in the code dealing
directly with the API so the rest of the code didn't have to concern
itself with this.